Improve SVM accuracy and add learning curve
In this commit, I have tuned C and gamma parameters of SVM, and it improves the accuracy, recall and f1 score little bit. The recall is now 90% which is good. At this stage, I want to see if adding more training data improves the model performance. Interestingly the learning shows that we need to add more data to get our model improved. It looks like accuracy can go up 95%.

In next step, I will collect more data and come back to learning curve to see the model performance.